+++ publish/sdk/src/exportedTypes.ts
@@ -37,7 +37,6 @@ export type {
   RegisterDevelopmentPluginResult,
   RepositoryNamespace,
   RetrievalCallbacks,
-  RetrievalNamespace,
   RetrievalOpts,
   RetrievalResult,
   RetrievalResultEntry,
@@ -73,8 +72,8 @@ export type {
   ColorPalette,
   DiagnosticsLogEvent,
   DiagnosticsLogEventData,
-  DownloadProgressUpdate,
   DownloadedModel,
+  DownloadProgressUpdate,
   EmbeddingLoadModelConfig,
   FileType,
   GPUSetting,
